seperate popularity_measure
